change datasource config

+++ spring-data-jpa-multisource/src/main/java/cn/springcamp/springdatajpa/multisource/other/data/OtherRepository.java
@@ -2,5 +2,5 @@ package cn.springcamp.springdatajpa.multisource.other.data;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OtherRepository extends JpaRepository<Other, Integer> {
+public interface OtherRepository extends JpaRepository<Other, Long> {
 }
